# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЭкоПульс API",
    "description": "API административной панели эколога/инспектора «ЭкоПульс» — триаж обращений о состоянии площадок ТКО.\n\n**Base URL:** `/api/v1`\n\n**Авторизация:** для защищённых эндпоинтов — Bearer JWT (`Authorization: Bearer <access_token>` из `POST /auth/login`). Приём фотоотчёта — заголовок `X-Intake-Token`. Отдача фото — публично.\n\n**Конверт ошибок:** `{ \"error\": { \"code\", \"message\", \"details\" } }`.\n\nРазделы, помеченные «вне мобильного API», в клиентскую документацию не входят.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/mno": {
      "get": {
        "tags": [
          "Карта и МНО"
        ],
        "summary": "List Mno",
        "description": "Реестр МНО с фильтрами region/synced/search + сортировкой.",
        "operationId": "list_mno_api_v1_mno_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "synced",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Synced"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MnoListItem"
                  },
                  "title": "Response List Mno Api V1 Mno Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Добавление нового МНО"
        ],
        "summary": "Create Mno",
        "description": "Добавление МНО вручную (synced=false, fgis_id=null, incidents=0).",
        "operationId": "create_mno_api_v1_mno_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MnoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mno/export": {
      "get": {
        "tags": [
          "Экспорт (вне мобильного API)"
        ],
        "summary": "Export Mno",
        "description": "Экспорт отфильтрованного реестра МНО в .xlsx.",
        "operationId": "export_mno_api_v1_mno_export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "synced",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Synced"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mno/sync": {
      "post": {
        "tags": [
          "Карта и МНО"
        ],
        "summary": "Sync All Mno",
        "description": "ЗАГЛУШКА: помечает все ещё-не-synced МНО как синхронизированные с ФГИС.",
        "operationId": "sync_all_mno_api_v1_mno_sync_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoSyncResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mno/{mno_id}/sync": {
      "post": {
        "tags": [
          "Карта и МНО"
        ],
        "summary": "Sync One Mno",
        "description": "ЗАГЛУШКА: синхронизирует одно МНО с ФГИС (без внешних запросов).",
        "operationId": "sync_one_mno_api_v1_mno__mno_id__sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mno_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mno Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mno/{mno_id}": {
      "get": {
        "tags": [
          "Карточка МНО"
        ],
        "summary": "Get Mno",
        "description": "Карточка МНО.",
        "operationId": "get_mno_api_v1_mno__mno_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mno_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mno Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/profile": {
      "patch": {
        "tags": [
          "Профиль пользователя"
        ],
        "summary": "Patch Profile",
        "description": "Обновляет Заявителя текущего пользователя.",
        "operationId": "patch_profile_api_v1_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/password": {
      "post": {
        "tags": [
          "Профиль пользователя"
        ],
        "summary": "Change Password",
        "description": "Смена собственного пароля (без проверки текущего, ТЗ §9.1).",
        "operationId": "change_password_api_v1_profile_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResult": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResult"
      },
      "MnoCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "coords": {
            "type": "string",
            "minLength": 1,
            "title": "Coords"
          },
          "reg": {
            "type": "string",
            "title": "Reg",
            "default": ""
          },
          "region_code": {
            "type": "string",
            "title": "Region Code",
            "default": ""
          },
          "city": {
            "type": "string",
            "title": "City",
            "default": ""
          },
          "address": {
            "type": "string",
            "title": "Address",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "coords"
        ],
        "title": "MnoCreate",
        "description": "Создание МНО вручную. Обязательны только name + coords.\n\nsynced=false, fgis_id=null, incidents=0 проставляются сервисом — НЕ из тела."
      },
      "MnoDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reg": {
            "type": "string",
            "title": "Reg"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "region_code": {
            "type": "string",
            "title": "Region Code"
          },
          "region_name": {
            "type": "string",
            "title": "Region Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "coords": {
            "type": "string",
            "title": "Coords"
          },
          "fgis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fgis Id"
          },
          "synced": {
            "type": "boolean",
            "title": "Synced"
          },
          "sync_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Date"
          },
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reg",
          "name",
          "region_code",
          "region_name",
          "city",
          "address",
          "coords",
          "synced",
          "incidents"
        ],
        "title": "MnoDetail",
        "description": "Карточка МНО — те же поля, что и в списке (контракт)."
      },
      "MnoListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reg": {
            "type": "string",
            "title": "Reg"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "region_code": {
            "type": "string",
            "title": "Region Code"
          },
          "region_name": {
            "type": "string",
            "title": "Region Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "coords": {
            "type": "string",
            "title": "Coords"
          },
          "fgis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fgis Id"
          },
          "synced": {
            "type": "boolean",
            "title": "Synced"
          },
          "sync_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Date"
          },
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reg",
          "name",
          "region_code",
          "region_name",
          "city",
          "address",
          "coords",
          "synced",
          "incidents"
        ],
        "title": "MnoListItem",
        "description": "Строка реестра МНО. region_name резолвится из справочника регионов."
      },
      "MnoSyncResult": {
        "properties": {
          "synced": {
            "type": "integer",
            "title": "Synced"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "synced",
          "total"
        ],
        "title": "MnoSyncResult",
        "description": "Итог ЗАГЛУШКИ синхронизации с ФГИС: сколько помечено / всего МНО."
      },
...
      "PasswordReset": {
        "properties": {
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "new_password"
        ],
        "title": "PasswordReset"
      },
...
      "ProfileUpdate": {
        "properties": {
          "fio": {
            "type": "string",
            "title": "Fio"
          }
        },
        "type": "object",
        "required": [
          "fio"
        ],
        "title": "ProfileUpdate"
      },
...
      "UserMe": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "fio": {
            "type": "string",
            "title": "Fio"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "invited"
            ],
            "title": "Status"
          },
          "is_superadmin": {
            "type": "boolean",
            "title": "Is Superadmin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "fio",
          "role",
          "status",
          "is_superadmin"
        ],
        "title": "UserMe"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}